producer - Generate random transaction id and amount

+++ src/main/java/com/volkanyungul/bank_account/producer/dto/Range.java
@@ -1,4 +1,4 @@
 package com.volkanyungul.bank_account.producer.dto;
 
-public record Range(double from, double to) {
+public record Range(long from, long to) {
 }
